remove size_t from exceptions and some keyboard fixes

+++ cpp_vk_lib/src/vk/error/exception.cpp
@@ -13,24 +13,24 @@ const char* common_exception::what() const noexcept
     return error_.what();
 }
 
-std::string common_exception::create(size_t id, const char* error_name, const char* arg)
+std::string common_exception::create(int32_t id, const char* error_name, const char* arg)
 {
     return fmt::format("[vk.exception.{}.{}]: {}", error_name, id, arg);
 }
 
-upload_error::upload_error(size_t id, const char* what_arg)
+upload_error::upload_error(int32_t id, const char* what_arg)
     : common_exception(create(id, "upload_error", what_arg))
 {}
 
-access_error::access_error(size_t id, const char* what_arg)
+access_error::access_error(int32_t id, const char* what_arg)
     : common_exception(create(id, "access_error", what_arg))
 {}
 
-invalid_parameter_error::invalid_parameter_error(size_t id, const char* what_arg)
+invalid_parameter_error::invalid_parameter_error(int32_t id, const char* what_arg)
     : common_exception(create(id, "invalid_parameter_error", what_arg))
 {}
 
-runtime_error::runtime_error(size_t id, const char* what_arg)
+runtime_error::runtime_error(int32_t id, const char* what_arg)
     : common_exception(create(id, "runtime_error", what_arg))
 {}
